requirements.txt
alembic==1.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.4
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.15.3
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
